requirements.txt
alembic==1.7.7
flask==1.1.2
flask_bootstrap==3.3.7.1
flask_login==0.5.0
flask_migrate==3.1.0
flask_sQLAlchemy==2.5.1
flask_WTF==1.0.0
numpy==1.21.2
paho_mqtt==1.6.1
pyserial==3.5
pyTelegramBotAPI==4.4.0
requests==2.27.1
SQLAlchemy==1.4.27
telegram_send==0.25
WTForms==3.0.1
jinja2<3.1.0